Add filterSourceRegex
Sane version of the regex matching in nixpkgs. Matches against the
whole subpath, as opposed to only one directory level at a time.

This is kind of alpha right now, as I only later discovered that
`filterSource` is “lazy” in what paths it checks, meaning it will add
all subpaths to the filter queue when you return `true` for a
directory.

Ideally, we’d accept a list of subdirectory matches in the future.