filterSourceGitignore: add filterSourceGitignoreWith
This more general function takes a list of gitignore lines instead of
a path to the file directly.
It will gain more optional arguments in the future and is thus
extensible.

There is a semantic change in that we use the `src` path as base
filter directory instead of the parent dir of the .gitignore, but that
should have been the case since the beginning really.