filterSourceGitignore: add globMap to *With
Now that we can add arguments, let’s make it possible to filter the
gitignore files in their most structured form.

Renames the `globs` argument to `gitignoreLine` which is more fitting.

This is not the last breaking change, in order to export these new
functions, the module has to be switched over to an attrset.